refactor check-upstream
to use shared classes from external modules and to place ALL checks under the FontBakeryChecks.py file.
(issue #1204)
