+++ FontBakeryChecks.py
@@ -28,6 +28,10 @@ from unidecode import unidecode
 from lxml.html import HTMLParser
 import plistlib
 from TargetFont import TargetFont
+from PiFont import PiFont,\
+                   PiFontSFD,\
+                   PiFontUfo,\
+                   PiFontFontTools
 from utils import get_name_string,\
                   check_bit_entry,\
                   getGlyph,\
@@ -39,7 +43,8 @@ from utils import get_name_string,\
                   glyphs_surface_area,\
                   save_FamilyProto_Message,\
                   parse_version_string,\
-                  font_key
+                  font_key,\
+                  assertExists
 from constants import IMPORTANT,\
                       CRITICAL,\
                       STYLE_NAMES,\
@@ -3344,3 +3349,173 @@ def check_regression_ttfauto_xheight_increase(fb, new_font, old_font, f):
   else:
     fb.ok("TTFAutohint --increase-x-height is the same as the previous "
           "release, %s" % (new_inc_xheight))
+
+
+###############################
+# Upstream Font Source checks #
+###############################
+
+def check_all_fonts_have_matching_glyphnames(fb, folder, directory):
+  fb.new_check(120, "Each font in family has matching glyph names?")
+  glyphs = None
+  failed = False
+  for f in directory.get_fonts():
+    try:
+      font = PiFont(os.path.join(folder, f))
+      if glyphs is None:
+        glyphs = font.get_glyphs()
+      elif glyphs != font.get_glyphs():
+        failed = True
+        fb.error(("Font '{}' has different glyphs in"
+                  " comparison to onther fonts"
+                  " in this family.").format(f))
+        break
+    except:
+      failed = True
+      fb.error("Failed to load font file: '{}'".format(f))
+
+  if failed is False:
+    fb.ok("All fonts in family have matching glyph names.")
+
+
+def check_glyphs_have_same_num_of_contours(fb, folder, directory):
+  fb.new_check("121", "Glyphs have same number of contours across family ?")
+  glyphs = {}
+  failed = False
+  for f in directory.get_fonts():
+    font = PiFont(os.path.join(folder, f))
+    for glyphcode, glyphname in font.get_glyphs():
+      contours = font.get_contours_count(glyphname)
+      if glyphcode in glyphs and glyphs[glyphcode] != contours:
+        failed = True
+        fb.error(("Number of contours of glyph '{}'"
+                  " does not match."
+                  " Expected {} contours, but actual is"
+                  " {} contours").format(glyphname,
+                                         glyphs[glyphcode],
+                                         contours))
+      glyphs[glyphcode] = contours
+  if failed is False:
+    fb.ok("Glyphs have same number of contours across family.")
+
+
+def check_glyphs_have_same_num_of_points(fb, folder, directory):
+  fb.new_check("122", "Glyphs have same number of points across family ?")
+  glyphs = {}
+  failed = False
+  for f in directory.get_fonts():
+    font = PiFont(os.path.join(folder, f))
+    for g, glyphname in font.get_glyphs():
+      points = font.get_points_count(glyphname)
+      if g in glyphs and glyphs[g] != points:
+        failed = True
+        fb.error(("Number of points of glyph '{}' does not match."
+                  " Expected {} points, but actual is"
+                  " {} points").format(glyphname,
+                                       glyphs[g],
+                                       points))
+      glyphs[g] = points
+  if failed is False:
+    fb.ok("Glyphs have same number of points across family.")
+
+
+def check_font_folder_contains_a_COPYRIGHT_file(fb, folder):
+  fb.new_check("123", "Does this font folder contain COPYRIGHT file ?")
+  assertExists(fb, folder, "COPYRIGHT.txt",
+               "Font folder lacks a copyright file at '{}'",
+               "Font folder contains COPYRIGHT.txt")
+
+
+def check_font_folder_contains_a_DESCRIPTION_file(fb, folder):
+  fb.new_check("124", "Does this font folder contain a DESCRIPTION file ?")
+  assertExists(fb, folder, "DESCRIPTION.en_us.html",
+               "Font folder lacks a description file at '{}'",
+               "Font folder should contain DESCRIPTION.en_us.html.")
+
+
+def check_font_folder_contains_licensing_files(fb, folder):
+  fb.new_check("125", "Does this font folder contain licensing files?")
+  assertExists(fb, folder, ["LICENSE.txt", "OFL.txt"],
+               "Font folder lacks licensing files at '{}'",
+               "Font folder should contain licensing files.")
+
+
+def check_font_folder_contains_a_FONTLOG_txt_file(fb, folder):
+  fb.new_check("126", "Font folder should contain FONTLOG.txt")
+  assertExists(fb, folder, "FONTLOG.txt",
+               "Font folder lacks a fontlog file at '{}'",
+               "Font folder should contain a 'FONTLOG.txt' file.")
+
+
+def check_repository_contains_METADATA_pb_file(fb, f):
+  fb.new_check("127", "Repository contains METADATA.pb file?")
+  fullpath = os.path.join(f, 'METADATA.pb')
+  if not os.path.exists(fullpath):
+    fb.error("File 'METADATA.pb' does not exist"
+             " in root of upstream repository")
+  else:
+    fb.ok("Repository contains METADATA.pb file.")
+
+
+def check_copyright_notice_is_consistent_across_family(fb, folder):
+  fb.new_check("128", "Copyright notice is consistent"
+                      " across all fonts in this family ?")
+
+  COPYRIGHT_REGEX = re.compile(r'Copyright.*?20\d{2}.*', re.U | re.I)
+  def grep_copyright_notice(contents):
+      match = COPYRIGHT_REGEX.search(contents)
+      if match:
+          return match.group(0).strip(',\r\n')
+      return
+
+  def lookup_copyright_notice(ufo_folder):
+      current_path = ufo_folder
+      try:
+          contents = open(os.path.join(ufo_folder,
+                                       'fontinfo.plist')).read()
+          copyright = grep_copyright_notice(contents)
+          if copyright:
+              return copyright
+      except (IOError, OSError):
+          pass
+
+      while os.path.realpath(self.operator.path) != current_path:
+          # look for all text files inside folder
+          # read contents from them and compare with copyright notice
+          # pattern
+          files = glob.glob(os.path.join(current_path, '*.txt'))
+          files += glob.glob(os.path.join(current_path, '*.ttx'))
+          for filename in files:
+              with open(os.path.join(current_path, filename)) as fp:
+                  match = COPYRIGHT_REGEX.search(fp.read())
+                  if not match:
+                      continue
+                  return match.group(0).strip(',\r\n')
+          current_path = os.path.join(current_path, '..')  # go up
+          current_path = os.path.realpath(current_path)
+      return
+
+  ufo_dirs = []
+  for root, dirs, files in os.walk(folder):
+      for d in dirs:
+          fullpath = os.path.join(root, d)
+          if os.path.splitext(fullpath)[1].lower() == '.ufo':
+              ufo_dirs.append(fullpath)
+  if len(ufo_dirs) == 0:
+      fb.skip("No UFO font file found.")
+  else:
+      failed = False
+      copyright = None
+      for ufo_folder in ufo_dirs:
+          current_notice = lookup_copyright_notice(ufo_folder)
+          if current_notice is None:
+              continue
+          if copyright is not None and current_notice != copyright:
+              failed = True
+              fb.error('"{}" != "{}"'.format(current_notice,
+                                             copyright))
+              break
+          copyright = current_notice
+      if failed is False:
+          fb.ok("Copyright notice is consistent"
+                " across all fonts in this family.")
